New:Be able to use stand-alone parser
