+++ .gitignore
@@ -132,3 +132,5 @@ dmypy.json
 # Custom
 ## IDEA
 .vscode
+## Codegen
+lark
