+++ jsonpath/lark.py
@@ -1,0 +1,7 @@
+try:
+    from lark import *
+    from lark.exceptions import *
+except ImportError:
+    import importlib.resources
+
+    exec(importlib.resources.read_text(__name__.split(".")[0], "lark"))
