+++ requirements-mini.txt
@@ -4,9 +4,6 @@ appdirs==1.4.3 \
 importlib-metadata==1.6.0; python_version < "3.8" \
     --hash=sha256:2a688cbaa90e0cc587f1df48bdc97a6eadccdcd9c35fb3f976a09e3b5016d90f \
     --hash=sha256:34513a8a0c4962bc66d35b359558fd8a5e10cd472d37aec5f66858addef32c1e
-lark-parser==0.8.5 \
-    --hash=sha256:e3ee17708e310de3a832a9beb448aa60ab6812f1bb283f72d473584c7566cc74 \
-    --hash=sha256:157879e04628e8e2bd57b72f3ba5642b3919137d5f7058612e95a38fb8761fa7
 six==1.14.0 \
     --hash=sha256:8f3cd2e254d8f793e7f3d6d9df77b92252b52637291d0f0da013c76ea2724b6c \
     --hash=sha256:236bdbdce46e6e6a3d61a337c0f8b763ca1e8717c03b369e87a7ec7ce1319c0a
