+++ setup.cfg
@@ -22,7 +22,10 @@ import_heading_firstparty=First Party Library
 import_heading_localfolder=Local Folder
 
 [tool:pytest]
-addopts = --doctest-modules
+addopts =
+    --doctest-glob jsonpath/core.py
+    --doctest-glob jsonpath/parser.py
+    --doctest-glob jsonpath/transformer.py
 
 [coverage:run]
 branch = true
