+++ tests/test_lark.py
@@ -8,13 +8,13 @@ from contextlib import nullcontext as does_not_raise
 import _pytest.python_api
 import pytest
 
-from lark import Lark, UnexpectedToken
-
 # First Party Library
 from jsonpath.core import JSONPathSyntaxError, JSONPathUndefinedFunctionError
+from jsonpath.lark import Lark, UnexpectedToken
 from jsonpath.parser import parse, parser
 
 
+@pytest.mark.xfail(raises=NameError)
 def test_no_conflict(caplog):
     with caplog.at_level(logging.DEBUG, logger="lark"):
         Lark.open(
